\name{data}
\alias{data}
\docType{data}
\title{
Edgar Anderson's Iris Data
}
\description{
\code{\link[datasets]{iris}} data
}
\usage{data}
\source{
Fisher, R. A. (1936) The use of multiple measurements in taxonomic problems. Annals of Eugenics, 7, Part II, 179-188.

The data were collected by Anderson, Edgar (1935). The irises of the Gaspe Peninsula, Bulletin of the American Iris Society, 59, 2-5.
}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.
}
\examples{
a=iris[1:50,1:4] # Iris data only for setosa 
}

