\name{hz-class}
\Rdversion{1.1}
\docType{class}
\alias{hz}
\alias{hz-class}


\title{Class \code{"hz"}}
\description{
An S4 class for Henze-Zirkler's Multivariate Normality Test
}

\section{Slots}{
  \describe{
    \item{\code{HZ}:}{stores the value of Henze-Zirkler statistic}
    \item{\code{p.value}:}{stores the p-value for the \code{HZ} test}
    \item{\code{dname}:}{stores the data set name}
    \item{\code{dataframe}:}{stores the data set}
  }
}
\author{
Selcuk Korkmaz, Dincer Goksuluk, Gokmen Zararsiz
}
