% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mardiaTest.R
\name{mardiaTest}
\alias{mardiaTest}
\title{Deprecated Function for Mardia's Multivariate Normality Test}
\usage{
mardiaTest(data, cov = TRUE, qqplot = FALSE)
}
\arguments{
\item{data}{a numeric matrix or data frame}

\item{cov}{If \code{TRUE} covariance matrix is normalized by \code{n}, if \code{FALSE} it is normalized by \code{n-1}}

\item{qqplot}{\code{TRUE} it creates chi-square Q-Q plot}
}
\description{
Please use 'mvn' function instead
}
