\name{MVN-package}
\alias{MVN-package}
\alias{MVN}
\docType{package}
\title{
Perform several multivariate normality tests.
}
\description{
The MVN library contain several multivariate normality tests such as Henze-Zirkler, Mardia and Royston.
}
\details{
\tabular{ll}{
Package: \tab MVN\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2014-03-02\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Selcuk Korkmaz, Dincer Goksuluk

Maintainer: Selcuk Korkmaz \email{selcuk.korkmaz@hacettepe.edu.tr}
}
