\name{royston-class}
\Rdversion{1.1}
\docType{class}
\alias{royston}
\alias{royston-class}


\title{Class \code{"royston"}}
\description{
An S4 class for Royston's Multivariate Normality Test
}

\section{Slots}{
  \describe{
    \item{\code{H}:}{stores the the value of Royston's H statistic}
    \item{\code{p.value}:}{stores the p-value for the Royston test}
    \item{\code{dname}:}{stores the data set name}
    \item{\code{dataframe}:}{stores the data set}
  }
}
\author{
Selcuk Korkmaz, Dincer Goksuluk, Gokmen Zararsiz
}
