% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roystonTest.R
\name{roystonTest}
\alias{roystonTest}
\title{Deprecated Function for Royston's Multivariate Normality Test}
\usage{
roystonTest(data, qqplot = FALSE)
}
\arguments{
\item{data}{a numeric matrix or data frame}

\item{qqplot}{\code{TRUE} it creates chi-square Q-Q plot}
}
\description{
Please use 'mvn' function instead
}
