\name{MVN-package}
\alias{MVN-package}
\alias{MVN}

\docType{package}
\title{
Performs several multivariate normality tests.
}
\description{
Assessing the assumption of multivariate normality is required by many parametric multivariate statistical methods, such as discriminant analysis, principal component analysis, MANOVA, etc. Here, we present an R package to asses multivariate normality. The MVN package contains three most widely used multivariate normality tests, including Mardia's, Henze-Zirkler's and Royston's multivariate normality tests. 
}
\details{
\tabular{ll}{
Package: \tab MVN\cr
Type: \tab Package\cr
License: \tab GPL (>= 2)\cr
}}
\author{
Selcuk Korkmaz, Dincer Goksuluk, Gokmen Zararsiz

Maintainer: Selcuk Korkmaz <selcuk.korkmaz@hacettepe.edu.tr>

}
