\name{MVN-package}
\alias{MVN-package}
\alias{MVN}
\docType{package}
\title{
Perform several multivariate normality tests.
}
\description{
The MVN library contain several multivariate normality tests such as Henze-Zirkler, Mardia and Royston.
}
\details{
\tabular{ll}{
Package: \tab MVN\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-02-04\cr
License: \tab GPL (>= 2)\cr
}
HZ.test(data)
mardia.test(data)
royston.test(data)
}
\author{
Selcuk KORKMAZ
Maintainer: Selcuk KORKMAZ \email{selcuk.korkmaz@hacettepe.edu.tr}
}
