\name{MVN-package}
\alias{MVN-package}
\alias{MVN}

\docType{package}
\title{
Performs several multivariate normality tests.
}
\description{
Performs multivariate normality tests and graphical approaches and implements multivariate outlier detection and univariate normality of marginal distributions through plots and tests.
}
\details{
\tabular{ll}{
Package: \tab MVN\cr
Type: \tab Package\cr
License: \tab GPL (>= 2)\cr
}}
\author{
Selcuk Korkmaz, Dincer Goksuluk, Gokmen Zararsiz

Maintainer: Selcuk Korkmaz <selcuk.korkmaz@hacettepe.edu.tr>

}
